+++ lib/i18n.d.ts
@@ -1,7 +1,7 @@
 import { Translations, ConvertMissingKey, Postprocess, Transforms, Preprocess } from './';
-export interface I18nContext<L extends string = string> {
+export interface I18nContext<L extends string = string, TG extends Translations<L> = Translations<L>> {
     locale: L;
-    globalTranslations?: Translations<L>;
+    globalTranslations: TG;
     convertMissingKey?: ConvertMissingKey;
     transforms?: Transforms;
     preprocess?: Preprocess;